+++ BE/src/main/java/com/newsainturtle/shadowmate/user/dto/response/ProfileResponse.java
@@ -1,0 +1,14 @@
+package com.newsainturtle.shadowmate.user.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ProfileResponse {
+    private String email;
+    private String nickname;
+    private String profileImage;
+    private String statusMessage;
+    private String plannerAccessScope;
+}
