+++ BE/src/main/java/com/newsainturtle/shadowmate/user/dto/response/SearchIntroductionResponse.java
@@ -1,4 +1,4 @@
-package com.newsainturtle.shadowmate.user.dto;
+package com.newsainturtle.shadowmate.user.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,6 +6,5 @@ import lombok.Getter;
 @Getter
 @Builder
 public class SearchIntroductionResponse {
-
     private String introduction;
 }
