+++ BE/src/main/java/com/newsainturtle/shadowmate/user/dto/response/UserResponse.java
@@ -1,4 +1,4 @@
-package com.newsainturtle.shadowmate.user.dto;
+package com.newsainturtle.shadowmate.user.dto.response;
 
 import com.newsainturtle.shadowmate.follow.enums.FollowStatus;
 import com.newsainturtle.shadowmate.user.enums.PlannerAccessScope;
@@ -8,18 +8,11 @@ import lombok.Getter;
 @Getter
 @Builder
 public class UserResponse {
-
     private Long userId;
-
     private String email;
-
     private String nickname;
-
     private String profileImage;
-
     private String statusMessage;
-
     private PlannerAccessScope plannerAccessScope;
-
     private FollowStatus isFollow;
 }
