+++ BE/src/main/java/com/newsainturtle/shadowmate/user/service/UserService.java
@@ -1,23 +1,20 @@
 package com.newsainturtle.shadowmate.user.service;
 
-import com.newsainturtle.shadowmate.user.dto.*;
+import com.newsainturtle.shadowmate.user.dto.request.UpdateIntroductionRequest;
+import com.newsainturtle.shadowmate.user.dto.request.UpdateUserRequest;
+import com.newsainturtle.shadowmate.user.dto.response.ProfileResponse;
+import com.newsainturtle.shadowmate.user.dto.response.SearchIntroductionResponse;
+import com.newsainturtle.shadowmate.user.dto.response.UserResponse;
 import com.newsainturtle.shadowmate.user.entity.User;
 import org.springframework.stereotype.Service;
 
 @Service
 public interface UserService {
-
     ProfileResponse getProfile(final Long userId);
-
     UserResponse searchNickname(final User user, final String nickname);
-
     SearchIntroductionResponse searchIntroduction(final Long userId);
-
     void updateUser(final Long userId, final UpdateUserRequest updateUserRequest);
-
     void updatePassword(final Long userId, final String oldPassword, final String newPassword);
-
     void updateIntroduction(final UpdateIntroductionRequest updateIntroductionRequest, final Long userId);
-
     void deleteUser(final User user);
 }
