+++ BE/src/test/java/com/newsainturtle/shadowmate/kh/user/UserControllerTest.java
@@ -7,10 +7,10 @@ import com.newsainturtle.shadowmate.auth.service.AuthServiceImpl;
 import com.newsainturtle.shadowmate.common.GlobalExceptionHandler;
 import com.newsainturtle.shadowmate.follow.enums.FollowStatus;
 import com.newsainturtle.shadowmate.user.controller.UserController;
-import com.newsainturtle.shadowmate.user.dto.UpdateIntroductionRequest;
-import com.newsainturtle.shadowmate.user.dto.UpdatePasswordRequest;
-import com.newsainturtle.shadowmate.user.dto.UpdateUserRequest;
-import com.newsainturtle.shadowmate.user.dto.UserResponse;
+import com.newsainturtle.shadowmate.user.dto.request.UpdateIntroductionRequest;
+import com.newsainturtle.shadowmate.user.dto.request.UpdatePasswordRequest;
+import com.newsainturtle.shadowmate.user.dto.request.UpdateUserRequest;
+import com.newsainturtle.shadowmate.user.dto.response.UserResponse;
 import com.newsainturtle.shadowmate.user.entity.User;
 import com.newsainturtle.shadowmate.user.enums.PlannerAccessScope;
 import com.newsainturtle.shadowmate.user.enums.SocialType;
